requirements.txt
Flask==2.3.3
Flask-Limiter==3.3.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0